// src/user/roles/role.enum.ts
export enum Role {
  USER = "user",
  ADMIN = "admin",
  CARRITO_ADMIN = "carrito_admin",
  CATEGORIA_ADMIN = "categoria_admin",
  CARRITO_REGALO_ADMIN = "carrito_regalo_admin",
  CATEGORIA_REGALO_ADMIN = "categoria_regalo_admin",
  CARRITO_LECTURA = "carrito_lectura",
  CATEGORIA_LECTURA = "categoria_lectura",
  CARRITO_REGALO_LECTURA = "carrito_regalo_lectura",
  CATEGORIA_REGALO_LECTURA = "categoria_regalo_lectura",
  CARRITO_ESCRITURA = "carrito_escritura",
  CATEGORIA_ESCRITURA = "categoria_escritura",
  CARRITO_REGALO_ESCRITURA = "carrito_regalo_escritura",
  CATEGORIA_REGALO_ESCRITURA = "categoria_regalo_escritura",
  CARRITO_ELIMINACION = "carrito_eliminacion",
  CATEGORIA_ELIMINACION = "categoria_eliminacion",
  CARRITO_REGALO_ELIMINACION = "carrito_regalo_eliminacion",
  CATEGORIA_REGALO_ELIMINACION = "categoria_regalo_eliminacion",
  USUARIO_ADMIN = "usuario_admin",
  USUARIO_LECTURA = "usuario_lectura",
  USUARIO_ESCRITURA = "usuario_escritura",
  USUARIO_ELIMINACION = "usuario_eliminacion",
  USUARIO_USUARIO_ADMIN = "usuario_usuario_admin",
  USUARIO_USUARIO_LECTURA = "usuario_usuario_lectura",
  USUARIO_USUARIO_ESCRITURA = "usuario_usuario_escritura",
  USUARIO_USUARIO_ELIMINACION = "usuario_usuario_eliminacion",
  USUARIO_CATEGORIA_ADMIN = "usuario_categoria_admin",
  USUARIO_CATEGORIA_LECTURA = "usuario_categoria_lectura",
  USUARIO_CATEGORIA_ESCRITURA = "usuario_categoria_escritura",
  USUARIO_CATEGORIA_ELIMINACION = "usuario_categoria_eliminacion",  
  RESENA_ADMIN = "resena_admin",
  RESENA_LECTURA = "resena_lectura",
  RESENA_ESCRITURA = "resena_escritura",
  RESENA_ELIMINACION = "resena_eliminacion",
  REGALO_RESENA_ADMIN = "regalo_resena_admin",
  REGALO_RESENA_LECTURA = "regalo_resena_lectura",
  REGALO_RESENA_ESCRITURA = "regalo_resena_escritura",
  REGALO_RESENA_ELIMINACION = "regalo_resena_eliminacion",
  USUARIO_RESENA_ADMIN = "usuario_resena_admin",
  USUARIO_RESENA_LECTURA = "usuario_resena_lectura",
  USUARIO_RESENA_ESCRITURA = "usuario_resena_escritura",
  USUARIO_RESENA_ELIMINACION = "usuario_resena_eliminacion",  
  USUARIO_WISHLIST_ADMIN = "usuario_wishlist_admin",
  USUARIO_WISHLIST_LECTURA = "usuario_wishlist_lectura",
  USUARIO_WISHLIST_ESCRITURA = "usuario_wishlist_escritura",
  USUARIO_WISHLIST_ELIMINACION = "usuario_wishlist_eliminacion",
  WISHLIST_ADMIN = "wishlist_admin",
  WISHLIST_LECTURA = "wishlist_lectura",
  WISHLIST_ESCRITURA = "wishlist_escritura",
  WISHLIST_ELIMINACION = "wishlist_eliminacion",
  REGALO_ADMIN = "regalo_admin",
  REGALO_ESCRITURA = "regalo_escritura",
  REGALO_LECTURA = "regalo_lectura",
  REGALO_ELIMINACION = "regalo_eliminacion",
  WISHLIST_REGALO_ADMIN = "wishlist_regalo_admin",
  WISHLIST_REGALO_ESCRITURA = "wishlist_regalo_escritura",
  WISHLIST_REGALO_LECTURA = "wishlist_regalo_lectura",
  WISHLIST_REGALO_ELIMINACION = "wishlist_regalo_eliminacion",
  REGALO_CATEGORIA_ADMIN = "regalo_categoria_admin",
  REGALO_CATEGORIA_ESCRITURA = "regalo_categoria_escritura",
  REGALO_CATEGORIA_LECTURA = "regalo_categoria_lectura",
  REGALO_CATEGORIA_ELIMINACION = "regalo_categoria_eliminacion",
  REGALO_TIENDA_ADMIN = "regalo_tienda_admin",
  REGALO_TIENDA_ESCRITURA = "regalo_tienda_escritura",
  REGALO_TIENDA_LECTURA = "regalo_tienda_lectura",
  REGALO_TIENDA_ELIMINACION = "regalo_tienda_eliminacion",

  TIENDA_ADMIN = "tienda_admin",
  TIENDA_LECTURA = "tienda_lectura",
  TIENDA_ESCRITURA = "tienda_escritura",
  TIENDA_ELIMINACION = "tienda_eliminacion",

  LOCALIZACION_ADMIN = "localizacion_admin",
  LOCALIZACION_LECTURA = "localizacion_lectura",
  LOCALIZACION_ESCRITURA = "localizacion_escritura",
  LOCALIZACION_ELIMINACION = "localizacion_eliminacion",

  TIENDA_REGALO_ADMIN = "tienda_regalo_admin",
  TIENDA_REGALO_LECTURA = "tienda_regalo_lectura",
  TIENDA_REGALO_ESCRITURA = "tienda_regalo_escritura",
  TIENDA_REGALO_ELIMINACION = "tienda_regalo_eliminacion",

  TIENDA_LOCALIZACION_ADMIN = "tienda_localizacion_admin",
  TIENDA_LOCALIZACION_LECTURA = "tienda_localizacion_lectura",
  TIENDA_LOCALIZACION_ESCRITURA = "tienda_localizacion_escritura",
  TIENDA_LOCALIZACION_ELIMINACION = "tienda_localizacion_eliminacion",

  USUARIO_DIRECCIONES_ADMIN = "usuario_direcciones_admin",
  USUARIO_DIRECCIONES_LECTURA = "usuario_direcciones_lectura",
  USUARIO_DIRECCIONES_ESCRITURA = "usuario_direcciones_escritura",
  USUARIO_DIRECCIONES_ELIMINACION = "usuario_direcciones_eliminacion",


}